feat：Support Native Animation LTPO



# message auto-generated for no-merge-commit merge:
merge wxjer_LTPO into master

feat：Support Native Animation LTPO

Created-by: wu_xujie1
Commit-by: wuxujie
Merged-by: huangyouhua
Description: <!-- 请在合并之前删除注释代码，以避免将其包含在GIT历史记录中 -->
# 描述
<!--
简要描述PR的目的，例如:
这条PR主要实现metro局域网内远程调试的功能. 
-->
对于native animation 的三类动画 frameAnimation/springAnimation/decayAnimation的位移动画支持 LTPO
# 变更日志
<!--
列出可包含在变更日志中的要点的主要变更。
不要列出内部更改。请以下列开头来写每个变更点：
- added
- fixed
- deprecated
- removed
如果是一条新增属性的变更，则以[added]开头，例如:
- [added] FlatList组件新增inverted属性
-->
-add：新增动画速率映射动态帧率，并native动画进行帧率投票
# 测试步骤
<!-- 
如何测试这条PR。例如：
1. 打开Tester工程.
2. 执行npm run start命令.
...
-->
1.应用这笔代码，测试frameAnimation/springAnimation/decayAnimation动态帧率变化

# 合入前自检
<!--
请在合入前根据以下条例进行自检
自检完成后在[ ]内填入"x", 例如：
- [x] 不涉及非兼容性变更；若涉及，已通过相应评审。
-->
- [x] 不涉及非兼容性变更；若涉及，已通过相应评审。
- [x] 不涉及性能或已进行性能测试且无性能劣化。
- [x] 符合对应的编码规范。
- [x] 不涉及文档更新，或已更新了文档。
- [x] 针对可测试性要求，已增加必要的自测用例、合理的日志记录或Trace信息。
- [x] 不存在非法的文件引入，包括图片和代码等
- [x] 本PR已经在77分支开启PR或不需要在77分支合入
  - 如需合入77分支，PR地址： 


See merge request: openharmony-sig/ohos_react_native!1106

+++ tester/harmony/react_native_openharmony/src/main/cpp/RNOHCorePackage/TurboModules/Animated/Drivers/AnimationDriver.h
@@ -31,7 +31,7 @@ class AnimationDriver {
     }
   };
 
-  virtual void runAnimationStep(uint64_t frameTimeNanos) = 0;
+  virtual void runAnimationStep(long long frameTimeNanos) = 0;
   virtual void resetConfig(folly::dynamic const& config);
 
   facebook::react::Tag getId() const;
